+++ resources/views/dashboard/brands/all.blade.php
@@ -36,9 +36,10 @@
 
 <div class="btn-group">
      
-    <a class="btn btn-warning" data-toggle="modal" data-target=".add_item"> 
-            <i class="fa fa-plus"></i>  @lang('site.add') </a> 
-     
+    @if(has_permission('brands_add'))
+        <a class="btn btn-warning" data-toggle="modal" data-target=".add_item"> 
+                <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+    @endif 
 </div>
 
 <br/> <br/>
@@ -68,11 +69,15 @@
                     {{__('site.view')}}  ({{count($item->models)}})  </a></td>                            
                 
                 <td>
-                    <a href="{{ route('admin.brand',$item->id) }}" class="btn btn-info btn-xs">
-                        <i class="fa fa-edit"></i> </a>
-    
+                    @if(has_permission('brands_edit'))
+                        <a href="{{ route('admin.brand',$item->id) }}" class="btn btn-info btn-xs">
+                            <i class="fa fa-edit"></i> </a>
+                    @endif
+
+                    @if(has_permission('brands_delete'))
                         <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs">
                             <i class="fa fa-trash"></i> </a>
+                    @endif
                 </td>
             </tr>
         @endforeach
