+++ resources/views/dashboard/models/all.blade.php
@@ -11,9 +11,10 @@
   
 <div class="btn-group">
      
-    <a class="btn btn-warning" data-toggle="modal" data-target=".add_item"> 
-            <i class="fa fa-plus"></i>  @lang('site.add') </a> 
-     
+    @if(has_permission('models_add'))
+        <a class="btn btn-warning" data-toggle="modal" data-target=".add_item"> 
+                <i class="fa fa-plus"></i>  @lang('site.add') </a> 
+    @endif 
 </div>
 
 <br/> <br/>
@@ -34,11 +35,15 @@
             <td>{{$item['name_'.my_lang()]}}</td>                            
             
             <td>
-                <a href="{{ route('admin.model',$item->id) }}" class="btn btn-info btn-xs">
-                    <i class="fa fa-edit"></i> </a>
+                @if(has_permission('models_edit'))
+                    <a href="{{ route('admin.model',$item->id) }}" class="btn btn-info btn-xs">
+                        <i class="fa fa-edit"></i> </a>
+                @endif
 
+                @if(has_permission('models_delete'))
                    <a onclick="deleteItem({{ $item->id }})" class="btn btn-danger btn-xs">
                        <i class="fa fa-trash"></i> </a>                       
+                @endif
             </td>
         </tr>
     @endforeach
